+++ test/TestableHttpClient.IntegrationTests/AssertingRequestsWithNFluent.cs
@@ -1,5 +1,6 @@
 ﻿namespace TestableHttpClient.IntegrationTests;
 
+[Obsolete("Testing deprecated warnings")]
 public class AssertingRequestsWithNFluent
 {
     [Fact]